day5/employeeapidocker/requirements.txt
Django>=4.2.0
requests>=2.31.0
djangorestframework>=3.14.0
psycopg2>=2.9.0
django-rest-swagger>=2.2.0
drf-yasg>=1.21.0
setuptools>=65.5.0
hvac>=0.12.0
spring-config-client>=0.1.0
python-dotenv>=1.0.0
elasticsearch
elasticsearch-dsl
django-elasticsearch-dsl
CMRESHandler
